style(c): apply correct formatting with clang-format

+++ implementations/c/lib/codec/tests/variable_length_encoded_u2le_test.c
@@ -10,44 +10,44 @@
 
 #define TEST_SET_SIZE 0xffffu
 
-uint16_t *test_nums;
-uint8_t *ul2_encoded;
+uint16_t* test_nums;
+uint8_t*  ul2_encoded;
 
-int _test_codec_variable_length_encoded_u2le_setup(void **state) {
+int _test_codec_variable_length_encoded_u2le_setup(void** state)
+{
   int status = 0;
 
-  test_nums = (uint16_t *)malloc(TEST_SET_SIZE * sizeof(uint16_t));
+  test_nums = (uint16_t*) malloc(TEST_SET_SIZE * sizeof(uint16_t));
   if (0 == test_nums) {
     status = OCKAM_ERROR_INTERFACE_CODEC;
     goto exit_block;
   }
 
-  ul2_encoded = (uint8_t *)malloc(TEST_SET_SIZE * sizeof(uint16_t));
+  ul2_encoded = (uint8_t*) malloc(TEST_SET_SIZE * sizeof(uint8_t));
   if (0 == ul2_encoded) {
     status = OCKAM_ERROR_INTERFACE_CODEC;
     goto exit_block;
   }
 
-  for (int i = 0; i <= TEST_SET_SIZE; ++i) {
-    test_nums[i] = i;
-  }
+  for (int i = 0; i <= TEST_SET_SIZE; ++i) { test_nums[i] = i; }
 
 exit_block:
   return status;
 }
 
-void _test_codec_variable_length_encoded_u2le(void **state) {
-  uint8_t *out = ul2_encoded;
-  uint8_t *_out = 0;
-  uint8_t *in = ul2_encoded;
-  uint8_t *in_end = 0;
-  uint16_t value = 0;
+void _test_codec_variable_length_encoded_u2le(void** state)
+{
+  uint8_t* out                = ul2_encoded;
+  uint8_t* _out               = 0;
+  uint8_t* in                 = ul2_encoded;
+  uint8_t* in_end             = 0;
+  uint16_t value              = 0;
   uint16_t ul2_encoded_length = 0;
-  int i = 0;
+  int      i                  = 0;
 
   for (i = 0; i <= TEST_SET_SIZE; ++i) {
     _out = out;
-    out = encode_variable_length_encoded_u2le(out, test_nums[i]);
+    out  = encode_variable_length_encoded_u2le(out, test_nums[i]);
     if (test_nums[i] & 0xC000u) {
       assert_null(out);
       out = _out;
@@ -55,7 +55,7 @@ void _test_codec_variable_length_encoded_u2le(void **state) {
   }
 
   ul2_encoded_length = out - ul2_encoded;
-  in_end = in + ul2_encoded_length;
+  in_end             = in + ul2_encoded_length;
 
   i = 0;
   while (in < in_end) {
@@ -66,7 +66,8 @@ void _test_codec_variable_length_encoded_u2le(void **state) {
   }
 }
 
-int _test_codec_variable_length_encoded_u2le_teardown(void **state) {
+int _test_codec_variable_length_encoded_u2le_teardown(void** state)
+{
   if (0 != ul2_encoded) free(ul2_encoded);
   if (0 != test_nums) free(test_nums);
